feat: configuration broadcasting is done outside hLLM engine
